+++ r8s/r8scli/group/application_licenses.py
@@ -58,11 +58,5 @@ def delete(application_id, force):
     Deletes RIGHTSIZER_LICENSES Application.
     """
     from r8scli.service.initializer import init_configuration
-    if force:
-        click.confirm(
-            f'Do you really want to completely '
-            f'delete application {application_id}?',
-            abort=True
-        )
     return init_configuration().application_licenses_delete(
         application_id=application_id, force=force)