+++ r8s/r8scli/group/job.py
@@ -41,9 +41,11 @@ def describe(job_id=None, job_name=None, limit=None):
               help='Processing end date. Format: "%Y-%m-%d" '
                    'Example: 2023-06-20. If not set, scan will be '
                    'limited by tomorrow\'s date.')
+@click.option('--force_rescan', '-fr', is_flag=True,
+              help='To rescan instances without recent metric updates.')
 @cli_response()
 def submit(application_id, parent_id, scan_tenants,
-           scan_from_date, scan_to_date):
+           scan_from_date, scan_to_date, force_rescan):
     """
     Submits a R8s job.
     """
@@ -56,7 +58,9 @@ def submit(application_id, parent_id, scan_tenants,
         parent_id=parent_id,
         scan_tenants=scan_tenants,
         scan_from_date=scan_from_date,
-        scan_to_date=scan_to_date)
+        scan_to_date=scan_to_date,
+        force_rescan=force_rescan
+    )
 
 
 @job.command(cls=ViewCommand, name='terminate')