+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.12.0] - 2025-02-06
+## [3.13.0] - 2025-05-09
 * Extend resource group support:
   * Add `r8s parent group` cli group:
     * `add_allowed_tags` - Adds specified tags to resource group configuration
@@ -12,6 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `add_resource_group` - Adds specified resource groups to configuration
     * `remove_resource_group` - Removes specified resource groups from configuration
 
+## [3.12.0] - 2025-03-14
+* Add ability to run Syndicate RightSizer instance with SaaS ModularSDK / S3
+* `r8s application delete` - Do not remove the application if the linked parents exist. if --force, both application and linked parents will be permanently deleted
+* `r8s license delete` - Fix issue on describing linked algorithms
+* Remove separate confirmation from `r8s application/parent delete --force` - not supported in Modular
+* Support separate secret managers for r8s/modular
+* `r8s job describe`: changed default limit to 10 jobs
+* `r8s job submit`: add optional `--force_rescan` parameter, to force Rightsizer to generate recommendation for instances without metric updates
+* Reason of tenant scan failure will now be saved in Job.tenant_fail_reason_map($tenant:$reason)
 
 ## [3.11.0] - 2024-09-17
 * Update onprem version of the service